requirements.txt
streamlit>=1.24.0
pandas>=1.5.3
plotly>=5.13.1
matplotlib>=3.7.1
ollama>=0.1.0
PyMuPDF==1.21.1  # Using a specific version known to work
python-dotenv>=1.0.0 